Implementa il calcolo delle collisioni
Corregge un bug dovuto all'uso di List.fill invece di List.tabulate. Il riuso degli stessi oggetti
di tipo Tile avrebbe causato problemi con la gestione dei GameObject contenuti al loro interno.
Sposta le implementazioni di MapEnhanced e CharacterEnhanced in GameHelpers per permetterne il riuso.
