+++ pacman-lib/src/main/scala/it/unibo/scalapacman/lib/model/Character.scala
@@ -2,7 +2,7 @@ package it.unibo.scalapacman.lib.model
 
 import it.unibo.scalapacman.lib.math.Point2D
 
-trait Character {
+trait Character extends GameObject {
   val position: Point2D
   val speed: Double
   val direction: Direction
