+++ pacman-lib/src/main/scala/it/unibo/scalapacman/lib/model/GameObject.scala
@@ -1,0 +1,3 @@
+package it.unibo.scalapacman.lib.model
+
+trait GameObject
